requirements.txt
fastapi>=0.78.0,<0.79.0
pydantic==1.9.1
uvicorn>==0.17.6,<0.18.6
psycopg2-binary==2.9.3
SQLAlchemy>=1.4.37,<1.5.37
Jinja2==3.1.2
python-dotenv~=0.20.0